backend/requirements.txt
flask
flask-cors
PyPDF2
python-dotenv
requests
ibm-watsonx-ai